prodev_backend/server/apps/leaderboard/serializers.py
# server/apps/leaderboard/serializers.py

from rest_framework import serializers
from .models import LeaderboardEntry


class LeaderboardEntrySerializer(serializers.ModelSerializer):
    class Meta:
        model = LeaderboardEntry
        fields = ["id", "user", "challenge", "score"]
